image-processor/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-vision==3.4.5
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0